+++ Specs/Scene/IonImageryProviderSpec.js
@@ -354,7 +354,7 @@ describe("Scene/IonImageryProvider", function () {
   it("createImageryProvider works with MAPBOX", function () {
     return testExternalImagery(
       "MAPBOX",
-      { url: "http://test.invalid", mapId: 1 },
+      { accessToken: "test-token", url: "http://test.invalid", mapId: 1 },
       MapboxImageryProvider
     );
   });